requirements.txt
requests
python-dotenv
websocket-client
x10-perpetual-client
eth-account